+++ packages/pix-color-scheme-switcher/src/components/PixColorSchemeSwitcher/pix-color-scheme-switcher.attributes.js
@@ -1,0 +1,8 @@
+// @ts-check
+
+/**
+ * Attribute change handlers for PixColorSchemeSwitcher.
+ *
+ * This component currently has no observed attributes.
+ */
+export default {};